+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - feat: Support Window.fps= (#30)
 - feat: dxopal server prints access log (#31)
 - feat: Default index.html avoids main.rb cache (#32)
+- feat: Add Image.hsl2rgb (#34)
 
 ## v1.4.4 (2020-01-21)